# docs/admin-notification-api.openapi.yaml
openapi: 3.0.3
info:
  title: NL2SQL Admin Notification API
  version: 1.0.0
  description: |
    Read-only integration API for an external admin backend to monitor background
    NL2SQL query jobs, poll completion notifications, and retrieve completed results.

    Background jobs are created by the chatbot. This API intentionally does not
    accept raw SQL submissions.

    Recommended polling flow:

    1. Keep `last_id` in the admin backend, initially `0`.
    2. Every 15 seconds call `GET /api/integration/notifications?after_id={last_id}`.
    3. Process returned items in ascending `id` order.
    4. Replace the stored cursor with the response `last_id`.
    5. If the response contains `limit` items, immediately poll again with the new
       cursor before waiting another 15 seconds.
    6. Use `unread_count` for the badge and mark an item read after the admin opens it.

    The current read state is global. Marking a notification read affects every admin
    client. The optional `recipient_id` identifies the chatbot requester, not the
    administrator viewing the notification.

    The status-only endpoint is public and returns no job details. Keep the API key
    required by every detailed endpoint in the admin application's backend. Do not
    embed it in public browser JavaScript.
servers:
  - url: "{baseUrl}"
    description: NL2SQL deployment
    variables:
      baseUrl:
        default: http://192.168.100.136:8000
        description: Replace with the reachable NL2SQL server URL.
security:
  - ApiKeyAuth: []
tags:
  - name: Notifications
    description: Poll and acknowledge background-query notifications.
  - name: Query jobs
    description: Inspect jobs, retrieve results, and request cancellation.
paths:
  /api/integration/notifications:
    get:
      tags: [Notifications]
      operationId: listNotifications
      summary: List notifications after a cursor
      description: |
        Returns notifications in ascending ID order. For incremental polling, omit
        `unread` and advance `after_id` using the returned `last_id`. Set `unread=true`
        when loading an unread-only view, not as the only durable event cursor.
      parameters:
        - $ref: "#/components/parameters/AfterId"
        - name: unread
          in: query
          required: false
          description: Return only notifications that have not been marked read.
          schema:
            type: boolean
            default: false
        - name: recipient_id
          in: query
          required: false
          description: Filter by the chatbot requester ID, usually an email or guest ID.
          schema:
            type: string
            maxLength: 320
          example: guest@example.com
        - name: limit
          in: query
          required: false
          description: Maximum number of notifications returned.
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
      responses:
        "200":
          description: Notification page returned successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationList"
              example:
                items:
                  - id: 12
                    job_id: 1a786a9a-c5f3-4ae5-a6f8-c65fc773496b
                    recipient_id: guest@example.com
                    type: query.completed
                    title: Truy vấn đã hoàn thành
                    message: Truy vấn trả về 10000 dòng và đã đạt giới hạn kết quả.
                    read_at: null
                    created_at: "2026-09-13T04:46:29.883208+00:00"
                unread_count: 1
                last_id: 12
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "503":
          $ref: "#/components/responses/Unavailable"

  /api/integration/notifications/{notification_id}/read:
    post:
      tags: [Notifications]
      operationId: markNotificationRead
      summary: Mark a notification as read
      description: No request body is required. Calling this repeatedly is safe.
      parameters:
        - $ref: "#/components/parameters/NotificationId"
      responses:
        "200":
          description: Notification is marked read.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarkReadResponse"
              example:
                ok: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "503":
          $ref: "#/components/responses/Unavailable"

  /api/integration/query-jobs:
    get:
      tags: [Query jobs]
      operationId: listQueryJobs
      summary: List background query jobs
      parameters:
        - name: status
          in: query
          required: false
          description: Filter jobs by lifecycle status.
          schema:
            $ref: "#/components/schemas/JobStatus"
        - name: requested_by
          in: query
          required: false
          description: Filter by the chatbot requester ID.
          schema:
            type: string
            maxLength: 320
          example: guest@example.com
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 50
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Job page returned successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryJobList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "503":
          $ref: "#/components/responses/Unavailable"

  /api/integration/query-jobs/{job_id}:
    get:
      tags: [Query jobs]
      operationId: getQueryJob
      summary: Get one background query job
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Job returned successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryJob"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/Unavailable"

  /api/integration/query-jobs/{job_id}/status:
    get:
      tags: [Query jobs]
      operationId: getPublicQueryJobStatus
      summary: Check whether a job is done
      description: |
        Public minimal endpoint. It requires no API key and returns only `done` or
        `not_done`. Queued and running jobs are `not_done`; succeeded, failed, and
        cancelled jobs are `done`. It intentionally does not reveal whether the job
        succeeded, its SQL, requester, error, or result.
      security: []
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Minimal completion status returned successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicJobStatus"
              examples:
                notDone:
                  value:
                    status: not_done
                done:
                  value:
                    status: done
        "404":
          $ref: "#/components/responses/NotFound"

  /api/integration/query-jobs/{job_id}/result:
    get:
      tags: [Query jobs]
      operationId: getQueryJobResult
      summary: Retrieve a completed query result
      description: |
        Available only while the job has status `succeeded` and before its result TTL
        expires. A successful response can be large; the admin backend should avoid
        forwarding all rows to the browser unless the user explicitly opens them.
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Result returned successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: The job has not completed successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                detail: Query job is not complete
        "410":
          description: The result expired or is no longer available.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                detail: Query result has expired
        "500":
          description: The stored result could not be read.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          $ref: "#/components/responses/Unavailable"

  /api/integration/query-jobs/{job_id}/cancel:
    post:
      tags: [Query jobs]
      operationId: cancelQueryJob
      summary: Request cancellation of a queued or running job
      description: |
        Queued jobs are cancelled immediately. Cancellation of a running Oracle call is
        cooperative: the current database call can finish or time out before the job is
        finally marked cancelled, and any result is discarded.
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "202":
          description: Cancellation request accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CancelResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: The job cannot be cancelled from its current status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          $ref: "#/components/responses/Unavailable"

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Value configured as `QUERY_JOB_API_KEY` on the NL2SQL server.

  parameters:
    AfterId:
      name: after_id
      in: query
      required: false
      description: Return only notifications with an ID greater than this cursor.
      schema:
        type: integer
        format: int64
        minimum: 0
        default: 0
    NotificationId:
      name: notification_id
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    JobId:
      name: job_id
      in: path
      required: true
      schema:
        type: string
        format: uuid

  responses:
    Unauthorized:
      description: Missing or invalid API key.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            detail: Missing or invalid query job API key
    NotFound:
      description: The requested notification or job does not exist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            detail: Query job not found
    ValidationError:
      description: One or more path or query parameters are invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
    Unavailable:
      description: Background jobs are disabled or the integration API key is not configured.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            detail: Configure QUERY_JOB_API_KEY

  schemas:
    JobStatus:
      type: string
      enum: [queued, running, succeeded, failed, cancelled]

    PublicJobStatus:
      type: object
      required: [status]
      additionalProperties: false
      properties:
        status:
          type: string
          enum: [done, not_done]

    NotificationType:
      type: string
      enum: [query.completed, query.failed, query.cancelled]

    Notification:
      type: object
      required:
        - id
        - job_id
        - recipient_id
        - type
        - title
        - message
        - read_at
        - created_at
      properties:
        id:
          type: integer
          format: int64
        job_id:
          type: string
          format: uuid
        recipient_id:
          type: string
          description: Chatbot user who requested the background query.
        type:
          $ref: "#/components/schemas/NotificationType"
        title:
          type: string
        message:
          type: string
        read_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time

    NotificationList:
      type: object
      required: [items, unread_count, last_id]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        unread_count:
          type: integer
          minimum: 0
          description: Total unread count after applying the optional recipient filter.
        last_id:
          type: integer
          format: int64
          minimum: 0
          description: Cursor to send as `after_id` on the next incremental request.

    QueryJob:
      type: object
      required:
        - id
        - conversation_id
        - requested_by
        - question
        - generated_sql
        - status
        - attempt_count
        - cancel_requested
        - row_limit_reached
        - created_at
        - updated_at
        - result_available
      properties:
        id:
          type: string
          format: uuid
        idempotency_key:
          type: string
          nullable: true
        conversation_id:
          type: string
        requested_by:
          type: string
        question:
          type: string
        generated_sql:
          type: string
          description: Validated read-only Oracle SQL generated by the chatbot.
        status:
          $ref: "#/components/schemas/JobStatus"
        attempt_count:
          type: integer
          minimum: 0
        cancel_requested:
          type: boolean
        row_count:
          type: integer
          minimum: 0
          nullable: true
        column_count:
          type: integer
          minimum: 0
          nullable: true
        row_limit_reached:
          type: boolean
          description: True when the stored result reached the configured maximum row count.
        result_expires_at:
          type: string
          format: date-time
          nullable: true
        error:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
        updated_at:
          type: string
          format: date-time
        result_available:
          type: boolean
        result_url:
          type: string
          nullable: true
          description: Relative result URL, present only while the result is available.

    QueryJobList:
      type: object
      required: [items, total, pagination]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/QueryJob"
        total:
          type: integer
          minimum: 0
        pagination:
          $ref: "#/components/schemas/Pagination"

    Pagination:
      type: object
      required: [limit, offset, returned, has_more, next_offset]
      properties:
        limit:
          type: integer
          minimum: 1
        offset:
          type: integer
          minimum: 0
        returned:
          type: integer
          minimum: 0
        has_more:
          type: boolean
        next_offset:
          type: integer
          minimum: 0
          nullable: true

    QueryResult:
      type: object
      required: [job_id, columns, row_count, row_limit_reached, rows]
      properties:
        job_id:
          type: string
          format: uuid
        columns:
          type: array
          items:
            type: string
        row_count:
          type: integer
          minimum: 0
        row_limit_reached:
          type: boolean
        rows:
          type: array
          items:
            type: object
            additionalProperties: true

    CancelResponse:
      type: object
      required: [accepted, job]
      properties:
        accepted:
          type: boolean
          enum: [true]
        job:
          $ref: "#/components/schemas/QueryJob"

    MarkReadResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
          enum: [true]

    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string

    ValidationErrorResponse:
      type: object
      properties:
        detail:
          type: array
          items:
            type: object
            additionalProperties: true